requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
